+++ sdk/poh-config/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "solana-poh-config"
+description = "Definitions of Solana's proof of history."
+documentation = "https://docs.rs/solana-poh-config"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+all-features = true
+rustdoc-args = ["--cfg=docsrs"]
+
+[dependencies]
+serde = { workspace = true, optional = true }
+serde_derive = { workspace = true, optional = true }
+solana-frozen-abi = { workspace = true, optional = true, features = [
+    "frozen-abi",
+] }
+solana-frozen-abi-macro = { workspace = true, optional = true, features = [
+    "frozen-abi",
+] }
+
+[dev-dependencies]
+solana-clock = { workspace = true }
+static_assertions = { workspace = true }
+
+[features]
+frozen-abi = [
+    "dep:solana-frozen-abi",
+    "dep:solana-frozen-abi-macro",
+]
+serde = ["dep:serde", "dep:serde_derive"]
+
+[lints]
+workspace = true
