+++ sdk/src/lib.rs
@@ -78,7 +78,6 @@ pub mod native_loader;
 pub mod net;
 pub mod nonce_account;
 pub mod offchain_message;
-pub mod poh_config;
 pub mod precompiles;
 pub mod program_utils;
 pub mod pubkey;
@@ -132,6 +131,8 @@ pub use solana_fee_structure as fee;
 pub use solana_inflation as inflation;
 #[deprecated(since = "2.1.0", note = "Use `solana-packet` crate instead")]
 pub use solana_packet as packet;
+#[deprecated(since = "2.2.0", note = "Use `solana-poh-config` crate instead")]
+pub use solana_poh_config as poh_config;
 #[deprecated(since = "2.1.0", note = "Use `solana-program-memory` crate instead")]
 pub use solana_program_memory as program_memory;
 #[deprecated(since = "2.1.0", note = "Use `solana_pubkey::pubkey` instead")]
